+++ config/config.go
@@ -17,7 +17,3 @@ func New(lang, loc string, trans []Translation) Config {
 func (c Config) LocationIsSet() bool {
 	return c.Location != ""
 }
-
-func (c Config) TranslationsAreSet() bool {
-	return len(c.Translations) > 0
-}
