--- option/option.go
@@ -1,10 +1,0 @@
-package option
-
-const (
-	Upcoming Option = "upcoming"
-	Online   Option = "online"
-	JustNow  Option = "justNow"
-	NoSuffix Option = "noSuffix"
-)
-
-type Option string
